% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{add.self.loops}
\alias{add.self.loops}
\title{add self loops on sink nodes}
\usage{
add.self.loops(W)
}
\arguments{
\item{W}{a transition probability matrix}
}
\value{
the updated probability matrix with self loops
}
\description{
given a transition probability matrix, this procedures adds
self loops to sink nodes (nodes without any exiting edge)
}
\examples{
W = matrix(c(0,0,0,0,0.5,0.5,1,0,0), nrow=3, byrow=TRUE)
add.self.loops(W)

}
