% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{all.to.sink}
\alias{all.to.sink}
\title{All nodes to sink}
\usage{
\method{all}{to.sink}(D)
}
\arguments{
\item{D}{a cpmc}
}
\description{
This function modifies a cpmc so that all nodes point to a sink node
(named 'sink_node')
}
\details{
this is an example to show the usage of apply
}
