% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{binary_radix_sort}
\alias{binary_radix_sort}
\title{Radix sort for a binary matrix}
\usage{
binary_radix_sort(mat)
}
\arguments{
\item{mat}{a binary matrix (of 0 and 1)}
}
\value{
the sorted matrix
}
\description{
Sort the rows of a binary matrix in ascending order
}
\examples{
require(Matrix)
m <- Matrix(c(1,1,0,1,0,0,0,1,1), sparse = TRUE, ncol = 3)
binary_radix_sort(m)

}
