% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{bool.prism.translate}
\alias{bool.prism.translate}
\title{Boolean to PRISM Boolean}
\usage{
bool.prism.translate(b)
}
\arguments{
\item{b}{boolean value}
}
\value{
translated value
}
\description{
Convert Boolean values in PRISM language
}
