% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{char.array.equal}
\alias{char.array.equal}
\title{Check if character arrays are equal}
\usage{
char.array.equal(a, b)
}
\arguments{
\item{a}{first char array}

\item{b}{second char array}
}
\value{
TRUE if the arrays are equal
}
\description{
arrays are sorted beforehand, the procedure is case sensitive
}
\examples{
a <- c("A","B","C")
b <- c("A","B","C")
c <- c("A","B","D")
char.array.equal(a,b)
char.array.equal(a,c)

}
