% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{compact.dataset.easy}
\alias{compact.dataset.easy}
\title{Compact dataset rows with plyr}
\usage{
compact.dataset.easy(dataset)
}
\arguments{
\item{dataset}{input dataset (mutational matrix)}
}
\value{
the compacted dataset (mutational matrix)
}
\description{
Use plyr library to count duplicate rows
and compact the dataset (mutational). The column
'freq' will contain the counts for each row.
}
\examples{
compact.dataset.easy(example.dataset())

}
