% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{compact_dataset_easy}
\alias{compact_dataset_easy}
\title{Compact dataset rows}
\usage{
compact_dataset_easy(dataset)
}
\arguments{
\item{dataset}{input dataset (mutational matrix)}
}
\value{
the compacted dataset (mutational matrix)
}
\description{
Count duplicate rows
and compact the dataset (mutational). The column
'freq' will contain the counts for each row.
}
\examples{
compact_dataset_easy(example_dataset())

}
