% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{corrplot_from_df}
\alias{corrplot_from_df}
\title{Dataframe based correlation plot}
\usage{
corrplot_from_df(df)
}
\arguments{
\item{df}{input dataset}
}
\value{
the computed correlation plot
}
\description{
Prepare correlation plot based on a dataframe
}
\examples{
corrplot_from_df(example_dataset())

}
