% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVisualization.R
\name{corrplot_from_mutational_matrix}
\alias{corrplot_from_mutational_matrix}
\title{Correlation plot from mutational matrix}
\usage{
corrplot_from_mutational_matrix(mutmatrix)
}
\arguments{
\item{mutmatrix}{input dataset}
}
\value{
the computed correlation plot
}
\description{
Prepare correlation plot based on a mutational matrix
}
\examples{
corrplot_from_mutational_matrix(example_dataset())

}
