% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{corrplot_genes}
\alias{corrplot_genes}
\title{Gene based correlation plot}
\usage{
corrplot_genes(df)
}
\arguments{
\item{df}{input dataset (mutational matrix)}
}
\value{
the computed correlation plot
}
\description{
Prepare correlation plot based on a
dataframe computed from genes' perspective
}
\examples{
corrplot_genes(example_dataset())

}
