% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{corrplot_samples}
\alias{corrplot_samples}
\title{Sample based correlation plot}
\usage{
corrplot_samples(df)
}
\arguments{
\item{df}{input dataset (mutational matrix)}
}
\value{
the computed correlation plot
}
\description{
Prepare correlation plot based on a dataframe
computed from samples' perspective
}
\examples{
corrplot_samples(example_dataset())

}
