% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{cpmc.apply.on.visit}
\alias{cpmc.apply.on.visit}
\title{Apply function to all cpmc nodes that are reachable from a source}
\usage{
cpmc.apply.on.visit(D, procedure, id = 0, label = NULL)
}
\arguments{
\item{D}{a cpmc}

\item{procedure}{a procedure that takes a list in input. This list is
such that the field D contains the cpmc, was.visited is a
Boolean vector containing the nodes that were already
visited by this procedure while current contains the id of the currently
visited node. The output of this procedure must be a named list
containing the updated version of was.visited and current.
(It is actually possible manipulate the DFS visit freely by
changing the output list accordingly)}

\item{id}{id of the source}

\item{label}{label of the source (used id is the lowest that matches)}
}
\description{
This auxiliary function applies a procedure to each node that is
reachable from a source of the cpmc.
Side effects of the procedure are kept during the execution.
}
