% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{cpmc.apply}
\alias{cpmc.apply}
\title{Apply function to all cpmc nodes}
\usage{
cpmc.apply(D, procedure)
}
\arguments{
\item{D}{a cpmc}

\item{procedure}{a procedure that takes a list in input. This list is
such that the field D contains the cpmc, was.visited is a
Boolean vector containing the nodes that were already
visited by this procedure while current contains the id of the currently
visited node. The output of this procedure must be a named list
containing the updated version of was.visited and current.}
}
\description{
This auxiliary function applies a procedure to each node of the cpmc.
Side effects of the procedure are kept during the execution.
}
