% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{cpmc.apply.treatment}
\alias{cpmc.apply.treatment}
\title{Apply treatment effects}
\usage{
cpmc.apply.treatment(D, treatment, treatments)
}
\arguments{
\item{D}{a cpmc}

\item{treatment}{treatment name}

\item{treatments}{list of all treatment (as extracted by the lexer-parser)}
}
\description{
This (experimental) procedure applies treatment effects
(sequential)
}
