% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{cpmc.normalize}
\alias{cpmc.normalize}
\title{Normalize exiting probabilities}
\usage{
cpmc.normalize(D)
}
\arguments{
\item{D}{a cpmc}
}
\description{
This procedure normalizes exiting probabilities for each node of this
cpmc. Nothing is done for a node if the sum of these probabilities is 0.
}
