% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{cpmc.prune}
\alias{cpmc.prune}
\title{Prune cpmc}
\usage{
cpmc.prune(D, id = 0, label = NULL)
}
\arguments{
\item{D}{a cpmc}

\item{id}{id of the source}

\item{label}{label of the source (alternatively)}
}
\description{
This procedure removes nodes not reachable form a source
}
