% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\docType{class}
\name{cpmc}
\alias{cpmc}
\title{Class to represent Cancer Progression Markov Chains CPMC}
\arguments{
\item{source}{source node id}

\item{destination}{destination node id}

\item{p}{transition probability}

\item{normalize}{set TRUE if the remaining probabilities
of exiting from the source must be normalized}

\item{unique}{if TRUE check before adding the node if it
has a unique label}

\item{id}{node id}

\item{label}{node label
(removes the node with this label and lowest id)}
}
\value{
the id of a label if found, 0 else

array of ids of neighbor nodes

the computed graph

TRUE if the operation was successful

TRUE if the operation was successful

the computed plot

number of nodes of this cpmc

TRUE if all tests are passed

a PRISM formatted string representing this cpmc
}
\description{
get node id (index in P) form label

get array of the ids of neighbors nodes

creates an igraph object representing the current cpmc

set probability value of a transition from a node to another

add a node to this cpmc

remove a node from its id or label

plot this cpmc

retrieve number of nodes of this cpmc

check if this is a valid cpmc

format this cpmc as a PRISM model
}
\details{
This representation is also well fit to encode any type of
labeled Markov chains
}
\section{Fields}{

\describe{
\item{\code{P}}{: a transition probability matrix}

\item{\code{labels}}{: list of character vector representing genotypes
(or a list of atomic proposition)}
}}


