% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{cpmc.remove.singleton}
\alias{cpmc.remove.singleton}
\title{Remove singleton}
\usage{
cpmc.remove.singleton(D)
}
\arguments{
\item{D}{a cpmc}
}
\description{
This procedure remove singleton (nodes that point only to thyself or nowhere)
from a cpmc
}
