% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw_ggraph}
\alias{draw_ggraph}
\title{ggplot graph output}
\usage{
draw_ggraph(out, digits = 4, ...)
}
\arguments{
\item{out}{the output object of CIMICE (es, from quick run)}

\item{digits}{precision for edges' weights}

\item{...}{other arguments for format_labels}
}
\value{
ggraph object representing g as described
}
\description{
Draws the output graph using ggplot
}
\examples{
draw_ggraph(quick_run(example_dataset()))

}
