% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw_visNetwork}
\alias{draw_visNetwork}
\title{VisNetwork graph output (default)}
\usage{
draw_visNetwork(g, W, labels)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}
}
\value{
visNetwork object representing g as described
}
\description{
Draws the output graph using VisNetwork
}
\examples{
require(dplyr)
preproc <- example_dataset() \%>\% dataset_preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
g <- graph_non_transitive_subset_topology(samples, labels)
W <- compute_weights_default(g, freqs)
draw_visNetwork(g,W,labels)

}
