% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{example.dataset}
\alias{example.dataset}
\title{Creates a simple example dataset}
\usage{
example.dataset()
}
\value{
a simple mutational matrix
}
\description{
Creates a simple example dataset
}
\examples{
example.dataset()

}
