% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{example_dataset}
\alias{example_dataset}
\title{Creates a simple example dataset}
\usage{
example_dataset()
}
\value{
a simple mutational matrix
}
\description{
Creates a simple example dataset
}
\examples{
example_dataset()

}
