% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{example_dataset_withFreqs}
\alias{example_dataset_withFreqs}
\title{Creates a simple example dataset with frequency column}
\usage{
example_dataset_withFreqs()
}
\value{
a simple mutational matrix
}
\description{
Creates a simple example dataset with frequency column
}
\examples{
example_dataset_withFreqs()

}
