% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugLexerParser.R
\name{find.gene}
\alias{find.gene}
\title{Prepare "find a gene" functions}
\usage{
find.gene(gene)
}
\arguments{
\item{gene}{gene to be searched for}
}
\value{
a function that looks for a specific gene (a string) in a list
        that takes that list in input. The output of that function is Boolean.
}
\description{
Creates a function to look for a specific gene in a list
}
\examples{
require(purrr)
gene <- "A"
label <- c("B","C","A")
(find.gene(gene))(label)

}
