% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{find.label}
\alias{find.label}
\title{Find id of node with the given label}
\usage{
find.label(label, labels)
}
\arguments{
\item{label}{a character array of labels or AP}

\item{labels}{a list of labels}
}
\value{
the first index for which labels matches. 0 if there is none.
}
\description{
Check if the given label is yet in use given a label and a list of labels
in which to search. Return the first index that matches.
}
\examples{
require(purrr)
a <- c("A","B","C")
b <- c("A","B","C")
c <- c("A","B","D")
d <- c(a,b,c)
find.label(c,d)

}
