% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{find.occ}
\alias{find.occ}
\title{Find occurrences}
\usage{
find.occ(some, all)
}
\arguments{
\item{some}{subset}

\item{all}{universe set}
}
\value{
an array representing the characteristic function of a subset of 'all'
}
\description{
Prepares a Boolean array representing the characteristic function
of a subset. 'some' and 'all' must be sortable. The order of the elements
in the output array follow the sorted order of the universe set 'all'
}
\examples{
a <- c("A","B","D")
b <- c("A","B","C", "F", "D")
find.occ(a,b)

}
