% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{fix_clonal_genotype}
\alias{fix_clonal_genotype}
\title{Manage Clonal genotype in data}
\usage{
fix_clonal_genotype(samples, freqs, labels)
}
\arguments{
\item{samples}{input dataset (mutational matrix) as matrix}

\item{freqs}{genotype frequencies (in the rows' order)}

\item{labels}{list of gene names (in the columns' order)}
}
\value{
a named list containing the fixed "samples", "freqs" and "labels"
}
\description{
Fix the absence of the clonal genotype in the data (if needed)
}
\examples{
require(dplyr)
# compact
compactedDataset <- compact_dataset_easy(example_dataset())
# remove frequencies column from the compacted dataset
samples <- as.matrix(compactedDataset \%>\% select(-freq))
# save genes' names
genes = colnames(samples)
# keep the information on frequencies for further analysis
freqs = as.matrix(compactedDataset \%>\% ungroup() \%>\% select(freq))
freqs = freqs/sum(freqs)
freqs = c(freqs,0)
# prepare node labels listing the mutated genes for each node
labels <- prepare_labels(samples, genes)
# fix clonal
fix <- fix_clonal_genotype(samples, freqs, labels)
samples = fix[["samples"]]
freqs = fix[["freqs"]]
labels = fix[["labels"]]

}
