% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{format_labels}
\alias{format_labels}
\title{Format labels for output object}
\usage{
format_labels(labels, max_col = 3, max_row = 3)
}
\arguments{
\item{labels}{a charachter vector of the labels to manage}

\item{max_col}{maximum number of identifiers in a single row for a label}

\item{max_row}{maximum number of rows of identifiers in a label}
}
\value{
the updated labels
}
\description{
Prepare labels based on multiple identifiers so that they do not excede
a certain size (if they do, a simple number is used)
}
\examples{
 
format_labels(c("A, B", "C, D, E"))

}
