% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{format_next_state}
\alias{format_next_state}
\title{Format labels in prism language}
\usage{
format_next_state(dest_id, label.table)
}
\arguments{
\item{dest_id}{destination node to be considered}

\item{lable.table}{matrix of the characteristic function of labels}
}
\value{
transition labels in prism language
}
\description{
Format property labels in the prism language. Labels are represented
by Boolean values, in the contest of tumor philogenetics, they represent
genotypes. Labels are updated by the transitions.
}
\examples{
cimice.out <- quick.run(example.dataset())
DTMC <- to.cpmc(cimice.out$topology, cimice.out$weights, cimice.out$labels)
lt <- make.label.table(DTMC$labels)
format_next_state(1,lt)

}
