% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{format_transition_probabilities}
\alias{format_transition_probabilities}
\title{format transition probabilities}
\usage{
format_transition_probabilities(W)
}
\arguments{
\item{W}{a transition probability matrix}
}
\value{
a character vector containing the transitions in prism syntax
}
\description{
convert a transition probability matrix into transitions written
in prism syntax
}
\examples{
W <- matrix(c(0.3,0.3,0.4,0,0.5,0.5,1,0,0), nrow=3, byrow=TRUE)
format_transition_probabilities(W)

}
