% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{format_transition_probabilities_with_labels}
\alias{format_transition_probabilities_with_labels}
\title{Format transitions in prism language}
\usage{
format_transition_probabilities_with_labels(P, lable.table)
}
\arguments{
\item{P}{transition probability matrix}

\item{lable.table}{matrix of the characteristic function of labels}
}
\value{
transitions in prism language
}
\description{
Format transitions in the prism language. Also manages property labels.
}
\examples{
cimice.out <- quick.run(example.dataset())
DTMC <- to.cpmc(cimice.out$topology, cimice.out$weights, cimice.out$labels)
lt <- make.label.table(DTMC$labels)
format_next_state(1,lt)

}
