% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{get_samples_lables_matches}
\alias{get_samples_lables_matches}
\title{Get associations between graph nodes and the samples of the original dataset}
\usage{
get_samples_lables_matches(gr, original_dataset)
}
\arguments{
\item{gr}{a networkD3 graph}

\item{original_dataset}{a dataset with row and column names (samples, genes)}
}
\value{
a character vector of comma separated sample identifiers. 
These lists consists of the identifiers of the samples that match a certain node in the graph.
(follows node's orderd in the graph).
}
\description{
Get associations between graph nodes and the samples of the original dataset
}
\examples{
require(dplyr)
require(networkD3)
out <- quick_run(example_dataset())
gr <- igraph_to_networkD3(out$topology)
gr$nodes$label <- unlist(out$labels, use.names=FALSE) 
get_samples_lables_matches(gr, example_dataset())

}
