% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{graph_non_transitive_subset_topology}
\alias{graph_non_transitive_subset_topology}
\title{Default preparation of graph topology}
\usage{
graph_non_transitive_subset_topology(samples, labels)
}
\arguments{
\item{samples}{mutational matrix}

\item{labels}{genotype labels}
}
\value{
a graph with the wanted topology
}
\description{
By default, CIMICE computes the relation between
genotypes using the subset relation.
For the following steps it is also important
that the transitive edges are removed.
}
\examples{
require(dplyr)
preproc <- example_dataset() \%>\% dataset_preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
graph_non_transitive_subset_topology(samples, labels)
}
