% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{is.not.unique.label}
\alias{is.not.unique.label}
\title{Check if the given label is yet in use}
\usage{
is.not.unique.label(label, labels)
}
\arguments{
\item{label}{a character array of labels or AP}

\item{labels}{a list of labels}
}
\value{
TRUE if the label is found
}
\description{
Check if the given label is yet in use given a label and a list of labels
in which to search
}
\examples{
a <- c("A","B","C")
b <- c("A","B","C")
c <- c("A","B","D")
d <- c(a,b,c)
is.not.unique.label(c,d)

}
