% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{labels_to_prism_properties}
\alias{labels_to_prism_properties}
\title{Prepare PRISM header}
\usage{
labels_to_prism_properties(lable.table, initial_state_description)
}
\arguments{
\item{lable.table}{matrix of the characteristic functions of the labels}

\item{initial_state_description}{list of labels that are true in the initial state
TODO: the initial state label could be retrieved directly by the labels of the cpmc}
}
\value{
translated value
}
\description{
Prepare PRISM header by defining state variable and lable variables.
}
