% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugLexerParser.R
\name{load.treatments}
\alias{load.treatments}
\title{Lex and parse treatment information}
\usage{
load.treatments(path)
}
\arguments{
\item{path}{path to treatment description file}
}
\value{
a data structure with treatment information
}
\description{
Creates a data structure as defined by 'parser' object that represents
the information included in a treatment description file.
Check vignettes for format.
}
\examples{
load.treatments('./inst/extdata/test.treatments')

}
