% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{make.dataset}
\alias{make.dataset}
\title{Dataset line by line construction: initialization}
\usage{
make.dataset(...)
}
\arguments{
\item{...}{gene names (do not use '"', the input
is automatically converted to strings)}
}
\value{
a mutational matrix without samples structured as dataframe
}
\description{
Initialize a dataset for "line by line" creation
}
\examples{
make.dataset(APC,P53,KRAS)

}
