% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{make.label.table}
\alias{make.label.table}
\title{Create table of the labels}
\usage{
make.label.table(labels)
}
\arguments{
\item{labels}{set of labels}
}
\value{
matrix of characteristic functions
}
\description{
Translate labels to a matrix of characteristic functions. The
characteristic functions follow the ascending order of sorted labels.
}
\examples{
l <- list(c("A","B"), list("C"), list("F", "D"))
make.label.table(l)

}
