% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{make_dataset}
\alias{make_dataset}
\title{Dataset line by line construction: initialization}
\usage{
make_dataset(...)
}
\arguments{
\item{...}{gene names (do not use '"', the input
is automatically converted to strings)}
}
\value{
a mutational matrix without samples structured as (sparse) matrix
}
\description{
Initialize a dataset for "line by line" creation
}
\examples{
make_dataset(APC,P53,KRAS)

}
