% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{quick_run}
\alias{quick_run}
\title{Run CIMICE defaults}
\usage{
quick_run(dataset, mode = "CAPRI")
}
\arguments{
\item{dataset}{a mutational matrix as a data frame}

\item{mode}{indicates the used input format. Must be either "CAPRI" or "CAPRIpop"}
}
\value{
a list object representing the graph computed by CIMICE with the
structure `list(topology = g, weights = W, labels = labels)`
}
\description{
This function executes CIMICE analysis on a dataset using default settings.
}
\examples{
quick_run(example_dataset())

}
