% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read}
\alias{read}
\title{Read a "CAPRI" file}
\usage{
read(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\value{
the described mutational matrix as a (sparse) matrix
}
\description{
Read a "CAPRI" formatted file, as read_CAPRI
}
\examples{
read(system.file("extdata", "example.CAPRI", package = "CIMICE", mustWork = TRUE))

}
