% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUtils.R
\name{read_CAPRI_string}
\alias{read_CAPRI_string}
\title{Read "CAPRI" file from a string}
\usage{
read_CAPRI_string(txt)
}
\arguments{
\item{txt}{string in valid "CAPRI" format}
}
\value{
the described mutational matrix as a dataframe
}
\description{
Read a "CAPRI" formatted file, from a text string
}
\examples{
read_CAPRI_string("
s\\\\g A B C D
S1 0 0 0 1
S2 1 0 0 0
S3 1 0 0 0
S4 1 0 0 1
S5 1 1 0 1
S6 1 1 0 1
S7 1 0 1 1
S8 1 1 0 1
")

}
