% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_CAPRIpop}
\alias{read_CAPRIpop}
\title{Read a "CAPRIpop" file}
\usage{
read_CAPRIpop(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\value{
a list containing the described mutational matrix as a (sparse) matrix
        and a list of the frequency of the genotypes
}
\description{
Read a "CAPRIpop" formatted file from the file system
}
\examples{
#          "pathToDataset/myDataset.CAPRI"
read_CAPRI(system.file("extdata", "example.CAPRIpop", package = "CIMICE", mustWork = TRUE))

}
