% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_matrix}
\alias{read_matrix}
\title{Read dataset from an R matrix}
\usage{
read_matrix(mat)
}
\arguments{
\item{mat}{a boolean mutational matrix}
}
\value{
the sparse mutational matrix to be used as CIMICE's input
}
\description{
also converts that matrix to a sparse matrix
}
\examples{
m <- matrix(c(0,0,1,1,0,1,1,1,1), ncol=3)
colnames(m) <- c("A","B","C")
rownames(m) <- c("S1", "S2", "S3")
read_matrix(m)
}
