% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{to.prism}
\alias{to.prism}
\title{to prism module}
\usage{
to.prism(g, W, labels)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}
}
\value{
a string representing the graph as a prism model
}
\description{
convert this graph to a prism model
}
\examples{
require(dplyr)
require(purrr)
require(igraph)
preproc <- example.dataset() \%>\% dataset.preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
g <- graph.non.transitive.subset.topology(samples, labels)
W <- compute.weights.default(g, freqs)
to.prism(g,W,labels)

}
