% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{to_dot}
\alias{to_dot}
\title{Dot graph output}
\usage{
to_dot(g, W, labels)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}
}
\value{
a string representing the graph in dot format
}
\description{
Represents this graph in dot format (a textual output format)
}
\examples{
require(dplyr)
require(purrr)
preproc <- example_dataset() \%>\% dataset_preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
g <- graph_non_transitive_subset_topology(samples, labels)
W <- compute_weights_default(g, freqs)
to_dot(g,W,labels)

}
