% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{compact_dataset}
\alias{compact_dataset}
\title{Compact dataset rows}
\usage{
compact_dataset(mutmatrix)
}
\arguments{
\item{mutmatrix}{input dataset (mutational matrix)}
}
\value{
the compacted dataset (mutational matrix)
}
\description{
Count duplicate rows
and compact the dataset (mutational). The column
'freq' will contain the counts for each row.
}
\examples{
compact_dataset(example_dataset())

}
