% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVisualization.R
\name{corrplot_genes}
\alias{corrplot_genes}
\title{Gene based correlation plot}
\usage{
corrplot_genes(mutmatrix)
}
\arguments{
\item{mutmatrix}{input dataset (mutational matrix)}
}
\value{
the computed correlation plot
}
\description{
Prepare a correlation plot computed from genes' perspective
using a mutational matrix
}
\examples{
corrplot_genes(example_dataset())

}
