% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVisualization.R
\name{corrplot_samples}
\alias{corrplot_samples}
\title{Sample based correlation plot}
\usage{
corrplot_samples(mutmatrix)
}
\arguments{
\item{mutmatrix}{input dataset (mutational matrix)}
}
\value{
the computed correlation plot
}
\description{
Prepare a correlation plot computed from samples' perspective
using a mutational matrix
}
\examples{
corrplot_samples(example_dataset())

}
