% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmc.R
\name{cpmc.prune}
\alias{cpmc.prune}
\title{Prune cpmc}
\usage{
cpmc.prune(D, id = 0, label = NULL)
}
\arguments{
\item{D}{a cpmc}
}
\description{
This procedure normalizes exiting probabilities for each node of this
cpmc. Nothing is done for a node if the sum of these probabilities is 0.
}
