% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw_ggraph}
\alias{draw_ggraph}
\title{ggplot graph output}
\usage{
draw_ggraph(g, W, labels, digits = 4)
}
\arguments{
\item{g}{graph to be drawn}

\item{W}{weights on edges}

\item{labels}{node labels}

\item{digits}{precision for edges' weights}
}
\value{
ggraph object representing g as described
}
\description{
Draws the output graph using ggplot
}
\examples{
require(dplyr)
preproc <- example_dataset() \%>\% dataset_preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
g <- graph_non_transitive_subset_topology(samples, labels)
W <- compute_weights_default(g, freqs)
draw_ggraph(g,W,labels,digit = 3)

}
