% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw_networkD3}
\alias{draw_networkD3}
\title{NetworkD3 graph output}
\usage{
draw_networkD3(out, ...)
}
\arguments{
\item{out}{the output object of CIMICE (es, from quick run)}

\item{...}{other arguments for format_labels}
}
\value{
networkD3 object representing g as described
}
\description{
Draws the output graph using networkD3
}
\examples{
draw_networkD3(quick_run(example_dataset()))

}
