% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{draw_visNetwork}
\alias{draw_visNetwork}
\title{VisNetwork graph output (default)}
\usage{
draw_visNetwork(out, ...)
}
\arguments{
\item{out}{the output object of CIMICE (es, from quick run)}

\item{...}{other arguments for format_labels}
}
\value{
visNetwork object representing g as described
}
\description{
Draws the output graph using VisNetwork
}
\examples{
draw_visNetwork(quick_run(example_dataset()))

}
