% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVisualization.R
\name{gene_mutations_hist}
\alias{gene_mutations_hist}
\title{Histogram of genes' frequencies}
\usage{
gene_mutations_hist(mutmatrix, binwidth = 1)
}
\arguments{
\item{mutmatrix}{input dataset (mutational matrix)}

\item{binwidth}{binwidth parameter for the histogram (as in ggplot)}
}
\value{
the newly created histogram
}
\description{
Create the histogram of the genes' mutational frequencies
}
\examples{
gene_mutations_hist(example_dataset(), binwidth = 10)

}
