% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{get_samples_lables_matches}
\alias{get_samples_lables_matches}
\title{Get associations between graph nodes and the samples of the original dataset}
\usage{
get_samples_lables_matches(gr, original_dataset)
}
\arguments{
\item{gr}{a networkD3 graph}

\item{original_dataset}{a dataset with row and column names (samples, genes)}
}
\value{
a character vector of comma separated sample identifiers. 
These lists consists of the identifiers of the samples that match a certain node in the graph.
(follows node's orderd in the graph).
}
\description{
Get associations between graph nodes and the samples of the original dataset
}
\examples{
require(dplyr)
require(networkD3)
preproc <- example_dataset() \%>\% dataset_preprocessing
samples <- preproc[["samples"]]
freqs   <- preproc[["freqs"]]
labels  <- preproc[["labels"]]
genes   <- preproc[["genes"]]
g <- graph_non_transitive_subset_topology(samples, labels)
W <- compute_weights_default(g, freqs)
gr <- igraph_to_networkD3(g)
get_samples_lables_matches(gr, example_dataset())

}
