% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputUtils.R
\name{to_dot}
\alias{to_dot}
\title{Dot graph output}
\usage{
to_dot(out, ...)
}
\arguments{
\item{out}{the output object of CIMICE (es, from quick run)}

\item{...}{other arguments for format_labels}
}
\value{
a string representing the graph in dot format
}
\description{
Represents this graph in dot format (a textual output format)
}
\examples{
to_dot(quick_run(example_dataset()))

}
